// class-room-frontend/src/app/api/model/absenceDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LocalTime } from './localTime';
import { EleveDto } from './eleveDto';
import { PeriodeDto } from './periodeDto';
import { ClasseEleveDto } from './classeEleveDto';


export interface AbsenceDto { 
    idAbsence?: number;
    eleve?: EleveDto;
    date?: string;
    heure?: string;
    motif?: string;
    periode?: PeriodeDto;
    classeEleve?: ClasseEleveDto;
}